bai 9: create controller in MVC
